Implementing loading stopwords from stopwords.txt file

+++ src/main/java/WordCountRunner.java
@@ -6,7 +6,7 @@ public class WordCountRunner {
         System.out.print("Enter text: ");
         Scanner scanner = new Scanner(System.in);
         try {
-            int count = wordCount.countText(scanner.nextLine());
+            long count = wordCount.countText(scanner.nextLine());
             System.out.println("Number of words: " + count);
         } catch (IllegalArgumentException e) {
             System.out.println("Error: "+ e.getMessage());
